refactor(state): extract monitor state and monitor logic into monitor.rs
