+++ crates/halley-wl/src/wm/fullscreen.rs
@@ -28,16 +28,16 @@ impl HalleyWlState {
     fn fullscreen_monitor_name(&self, node_id: NodeId, output: Option<&WlOutput>) -> String {
         output
             .and_then(|requested_output| {
-                self.outputs.iter().find_map(|(name, output)| {
+                self.monitor_state.outputs.iter().find_map(|(name, output)| {
                     output.owns(requested_output).then_some(name.clone())
                 })
             })
-            .or_else(|| self.node_monitor.get(&node_id).cloned())
-            .unwrap_or_else(|| self.current_monitor.clone())
+            .or_else(|| self.monitor_state.node_monitor.get(&node_id).cloned())
+            .unwrap_or_else(|| self.monitor_state.current_monitor.clone())
     }
 
     fn fullscreen_monitor_view(&self, monitor_name: &str) -> (Vec2, Vec2) {
-        self.monitors
+        self.monitor_state.monitors
             .get(monitor_name)
             .map(|monitor| (monitor.viewport.center, monitor.viewport.size))
             .unwrap_or((self.viewport.center, self.viewport.size))
@@ -45,11 +45,11 @@ impl HalleyWlState {
 
 
     fn reset_monitor_zoom_once(&mut self, monitor_name: &str) {
-        if let Some(monitor) = self.monitors.get_mut(monitor_name) {
+        if let Some(monitor) = self.monitor_state.monitors.get_mut(monitor_name) {
             monitor.zoom_ref_size = monitor.viewport.size;
             monitor.camera_target_view_size = monitor.viewport.size;
         }
-        if self.current_monitor == monitor_name {
+        if self.monitor_state.current_monitor == monitor_name {
             self.zoom_ref_size = self.viewport.size;
             self.camera_target_view_size = self.viewport.size;
         }
@@ -75,7 +75,7 @@ impl HalleyWlState {
     }
 
     fn fullscreen_target_size_for(&self, monitor_name: &str) -> (i32, i32) {
-        self.outputs
+        self.monitor_state.outputs
             .get(monitor_name)
             .and_then(|output| output.current_mode())
             .map(|mode| (mode.size.w, mode.size.h))
@@ -357,7 +357,7 @@ impl HalleyWlState {
                     && n.kind == halley_core::field::NodeKind::Surface
                     && self.field.is_visible(id)
                     && self
-                        .node_monitor
+                        .monitor_state.node_monitor
                         .get(&id)
                         .is_none_or(|m| m == &monitor_name)
                     && self.node_intersects_monitor_viewport(id, monitor_name.as_str()))
@@ -495,10 +495,10 @@ impl HalleyWlState {
                 // it was displaced for is still active — i.e. the monitor it belongs
                 // to still has an active fullscreen session.
                 let node_monitor = self
-                    .node_monitor
+                    .monitor_state.node_monitor
                     .get(&id)
                     .cloned()
-                    .unwrap_or_else(|| self.current_monitor.clone());
+                    .unwrap_or_else(|| self.monitor_state.current_monitor.clone());
                 let displaced_for_active = self
                     .fullscreen_active_node
                     .contains_key(&node_monitor);
